Renamed Arrangements to HearingSupport
Minor refactoring due to design content change.
